requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multidict==6.4.3
narwhals==1.35.0
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
realtime==2.4.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.11.3
streamlit==1.44.1
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websockets==14.2
yarl==1.20.0
